requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
Pillow==10.2.0 
protobuf==4.25.1
python-docx==0.8.11
PyPDF2==3.0.1
gTTS==2.3.1
docx==0.8.11
